remove old 0.1.0beta documentation
It is getting confusing and time consuming to maintain both versions. A
lot of links are broken in `/docs/0.1.0beta` anyways, because of issues
described in #12.

We should consider `0.1.1beta` to now be the public release (`0.2.0`)
documentation, since the documentation actually written for `0.1.1beta` was
overridden by all the recent changes.