fix(plugins) remove the "open an issue" button
We are reevaluating our strategy for community management.

For now we want to leave GitHub issues for *bugs*, not general
questions or suggestions. At the same time the Gitter chat lacks
permanency and the mailing list visibility. While we decide what
to do, we've decided to just hide the button.

From #537