Enables header template to read passed parameters
This should be backward compatible, even as it alters a
common include file. NOTE: breadcrumb handling is not
yet sorted. I cannot yet figure out how that was
implemented. I think there is non-working code. Also
default icons are not working yet. We need a way to
determine if the image file exists or not; this is
likely going to be JavaScript in the change that will
add styling and JS.