docs(plugins-dev) remove mentions of `hooks.lua` (#988)
Removed in Kong 0.11, the old `hooks.lua` module is irrelevant to
subsequent versions of the plugins development guide. See:

https://github.com/Kong/kong/blob/master/CHANGELOG.md#0110---20170816

We keep the note on how the new process is similar to pre-0.11 in 0.11
and 0.12 versions, but remove this note from the 0.13 version.

The 0.14 version already got rid of mentions of the `hooks.lua` module.