Alter criteria for IDing versioned content
This code was checking for '.x' in directory names to identify
the directory as containing versoned content. Though there are
currently no other sibling directories at that tier (app/),
looking fora mere '.' would suffice, but replacing with a
regular expression that looks for values like '1.0' or '1.0.1'
or '1.0.x' but not 'about' or '_includes'.