fix(autodoc) replace resty with luajit for cli-generation
The resty executable has trouble dealing with `popen`. On
this particular case, it works on Linux but it throws a
`Interrupted system call` error on mac, when reading the last
line of each popen.

This PR replaces `resty` with `luajit` when generating the
CLI docs, which involve a call to `popen`.

Related with https://github.com/openresty/resty-cli/issues/35