docs(headers) change title to "introduction" (#1071)
Since ToC is automatically generated and the title is provided in the document header, it is redundant to have the title again in the first section. That section title will be "Introduction", and will be an h2 instead of an h1 to avoid all other sections being nested inside it.