feat(autodoc) implement autodoc for configuration file
This autodoc module parses the information contained in
Kong's config file, merges that with the `autodoc-conf/data.lua` file
and generates a new `configuration.md` file. The bulk of the work
happens in parser.lua, mostly due to the way the docs file is written,
with loose rules about how to indent / space paragraphs in descriptions.