feat(jekyll) tabs plugin
This adds a new Jekyll plugin which allows rendering tabs by using
Liquid tags:

```
{% tabs %}
{% tab This is Tab One %}

* This is the content of Tab One
* Tab content is expected to be markdown

{% tab This is Tab Two %}

And this is the content of Tab Two.

    ``` lua
    print("Source code is also allowed")
    ```

{% endtabs %}
```

Since the CSS for displaying tabs is not ready yet, for now this plugin
will render the tabs one after the other, like so:

```
** This is Tab One **

* This is the content of Tab One
* Tab content is expected to be markdown

** This is Tab Two **

And this is the content of Tab Two.

    ``` lua
    print("Source code is also allowed")
    ```
```

Once the CSS is available we can change every place by modifying the
tabs plugin itself.