docs(1.1.x/admin-api) add DB-less Admin API docs
Adds a new page for Admin API in DB-less Mode (which includes `/config`, and
where all entity endpoints only contain GET operations except for the healthcheck
setters).

The sidebar still points to the regular Admin API, but I added a pointer to DB-less
Mode at the top of the regular Admin API docs and vice-versa. This is a stopgap
measure before we do a more thorough reorganization of the Admin API reference docs.