docs(35/migrations) cover upgrade paths from Core (#1261)
Add upgrade instructions for users going from Kong Core to Kong Enterprise 0.35.

The 0.35 upgrade viability checker relies on either the presence of all 0.34 migrations or the absence of the schema_migrations table, which 1.0 removes. Because of this, 0.14 and 0.15 fail 0.35's upgrade check. These versions cannot be upgraded to 0.34, as 0.34 is based on Core 0.13. As such, the only viable upgrade path is to first upgrade them to 1.0.