docs(plugins) document how to use plugins in DB-less mode (#1323)
* chore(jekyll) restrict Jekyll version in Gemfile

Without this change a lot of different versions of Jekyll can be
installed when you run `bundle install` (given that we don't have a
Gemfile.lock commited in the repo). This change restricts how "new"
these versions can be, so that the plugins we create remain compatible.

* feat(jekyll) tabs plugin

This adds a new Jekyll plugin which allows rendering tabs by using
Liquid tags:

```
{% tabs %}
{% tab This is Tab One %}

* This is the content of Tab One
* Tab content is expected to be markdown

{% tab This is Tab Two %}

And this is the content of Tab Two.

    ``` lua
    print("Source code is also allowed")
    ```

{% endtabs %}
```

Since the CSS for displaying tabs is not ready yet, for now this plugin
will render the tabs one after the other, like so:

```
** This is Tab One **

* This is the content of Tab One
* Tab content is expected to be markdown

** This is Tab Two **

And this is the content of Tab Two.

    ``` lua
    print("Source code is also allowed")
    ```
```

Once the CSS is available we can change every place by modifying the
tabs plugin itself.

* feat(plugins) add db-less tabs/sections to plugin layout

* style(request-transformer) fix table format

The markdown parser was not rendering the tables with html inside as tables, rendering them incorrectly.

This change replaces some of those markdown tables with their html equivalents. It also reformats the remainig markdown tables by adding vertical bars to the sides, which seems to avoid further "confusion".

* docs(plugins) add db-less hand-made instructions to all open source plugins