Reference Routes and Services throughout current Enterprise tutorial (#1108)
* fix(getting-started) fix formatting for Add Svc

Remove unecessary indent that broke formatting for curl examples in the 0.13
Getting Started > Configuring a Service article.

* fix(getting-started) rename and redirect Add API

Move Enterprise Adding your API article to Configuring a Service to bring it
in line with core docs for 0.33 and 0.34. Add redirect from old locations.

* fix(getting-started) update next step in quickstep

* fix(getting-started) update Service copy

Update the Enterprise Getting Started > Configuring a Service copy to
refer to Routes and Services. Content mostly copied from 0.14 docs, but
rearranged a bit to fit the Enterprise docs' structure.

* fix(getting-started) update tutorial nav titles