docs(oidc) use sub for Okta consumer_claim
Change the consumer_claim from oid to sub for Okta. The original source
for this doc was the Azure guide, but Azure is the odd IDP out whose sub
claim values can't easily be used with our OIDC implementation, since
there's no way to know the correct value in advance of a sign-in
attempt.