docs(1.4.x/admin) generate lists of possible values from `one_of`
Our lists of accepted protocols in the service entity was outdated in the
1.4.x documentation.

Instead of fixing it once and risking getting this kind of thing out-of-date
again, we now generate the list of accepted values in the documentation, the
same way we do for presenting default values.

A companion PR was sent to the https://github.com/kong/kong
repo with the updated script.