Updated 1301 changelog, removed alpha tag
Removed "alpha" tag for license report feature in 1.3.0.1 changelog, as it's not an alpha feature, and we now have a topic in the docs. https://docs.konghq.com/enterprise/1.3-x/deployment/licenses/report/. While the topic applies to 1.3.0.2, the 1.3.0.1 license report info is valid and works, so not considered alpha. Our only hesitation is how Kong helps the user to run this report, and how we track info.