chore(k4k8s) add secrets for both Kong registries
Add instructions to create a secret for the
kong-enterprise-edition-docker registry in addition to the existing
kong-enterprise-k8s secret. We have a mix of customers using either
image in practice, and it's not immediately obvious that they require
separate secrets.

Several other changes in content to align the docs site example more
closely with the example in https://github.com/Kong/charts/pull/65