fix(k4k8s) add vault-auth to K4K8S changelog
Add vault-auth to the list of included plugins in the original K4K8S
release. vault-auth is present in the build configuration. The
versioning for that configuration is a bit odd, but the commit time
indicates it was in place before the initial release.

Alphabetize plugin list to make it easier to find a specific plugin in
the list.