updates for aws_key, aws_secret (#2205)
fix typo of aws.key and aws.secret to aws_key and aws_secret. also added to aws_key and aws_secret param reference that if aws_key and aws_secret are not set, it'll use EC2 instance's IAM role. Although that's mentioned in the "Notes" section, that's easy to miss and would be something that a user would expect to find when looking at aws_key and aws_secret.