docs(kic) use annotations for HTTPS redirect
Replace the KongIngress override binding with the equivalent dedicated
annotations in the ingress controller HTTPS redirect guide. While the
KongIngress method is still supported, the dedicated annotations are
more modern, and are supported in all KIC versions we cover on the docs
site.

Additionally, use a 301 rather than a 302. HTTP->HTTPS redirects are
typically permanent, so a 301 is more appropriate.