fix(k8s) correct invalid Deployment apiVersion (#2680)
* fix(k8s) correct invalid Deployment apiVersion

Pull fallback-svc example out of bit.ly/fallback-svc and correct the
apiVersion for its Deployment. The original link was a specific gist
version and can't be modified without creating a new link. The example
here is short enough that we can pull it out.

* chore(k8s) remove reference to HTTPBin in fallback

This guide uses an application somewhat similar to HTTPBin, but it's not
actually HTTPBin.