docs(ldap-auth-advanced) add more notes
Referrals are not supported in the plugin. One way to work around this is
to hit the LDAP Global Catalog instead, which is usually listenning on a
different port than the default `389`. This way, referrals don't get sent
back to the plugin.

The plugin doesn’t authenticate users (allow/deny requests) based on group
membership. For example:
- if the user is member of an LDAP group, the request will be allowed -> 200 OK
- if the user is _not_ member of a group, the request will still be allowed -> 200 OK