[DOCU-1437]Updates Immunity and StatsD exporter locations (#2821)
* Updates Immunity and StatsD exporter locations

With Bintray's sunsetting, we have had to move our assets to other
locations. The latest versions of Immunity (3.0.0 and 4.0.0) and the
StatsD exporter (0.3.1, 0.3.0, and 0.2) have been added to Docker Hub
and no longer require credentials to download.

* Apply suggestions from code review

Co-authored-by: lena-larionova <54370747+lena-larionova@users.noreply.github.com>

* Updates other versions with Lena's suggestions

* Adds additional wording updates based on moving assets to Docker Hub

* updates v2.2.x immunity install doc by replacing Bintray with Docker Hub

* Creates includes for statsd and immunity download

This commit includes the creation of two includes one for the statsd
install and one for immunity install. The include has conditions for
different versions.

* adds newline to end of new include files

* fix includes

Co-authored-by: lena-larionova <54370747+lena-larionova@users.noreply.github.com>
Co-authored-by: lena.larionova <yelena.larionova@gmail.com>