[DOCU-1460]Combines ee and ce plugin compatibility lists (#621)
* Combines ee and ce plugin compatibility lists

The Plugin Compatibility page includes an ee and ce tab and both have
a list of the different modes that can be used to deploy Kong. Both
Gateways can be deployed the same way and we want to make it clearer.
Particularly there should be no difference between them with db-less
mode.

* Combines ee and ce plugin compatibility lists

The Plugin Compatibility page includes an ee and ce tab and both have
a list of the different modes that can be used to deploy Kong. Both
Gateways can be deployed the same way and we want to make it clearer.
Particularly there should be no difference between them with db-less
mode.

* Adds Mocking plugin to enterprise plugin compatibility list and updates DB-less table heading