Updates include files and fixes links
The netlify builds were breaking bc of two broken links: one was
a mix up in a variable and one was a misplacement of a file. I moved
the file to the expected location and fixed the variable. Also, the
version variable needed to be updated in the includes files for
versions 2.x so I updated them.