[DOCU-1503] Clean up straggling 2.3 content in Gateway EE 2.4 (#2872)
* update 2.4 install version variable; update docker note

* add version to immunity table

* fix statsd version include

* straggling 2.3 mentions

* Updates include files and fixes links

The netlify builds were breaking bc of two broken links: one was
a mix up in a variable and one was a misplacement of a file. I moved
the file to the expected location and fixed the variable. Also, the
version variable needed to be updated in the includes files for
versions 2.x so I updated them.

* Adds update to v1.3 for Immunity with bintray sunsetting

Co-authored-by: Heather Cloward <heathercloward@gmail.com>