Updates style-guide, markdown-rules, contributing
This commit adds the contributing folder to the jekyll.yml, and
kong_versions.yml files to prevent local build errors. It also removes
the Line Numbers section from the markdown-rules.md files
because it was also breaking my local builds bc of an issue with
the highlight liquid tag. Futhermore, this comit formats some of the
content in the style-guide.md file.