Adds incudes, hones headings, and fixes notes
Per Gayle's review, instead of having links to the instructions
for installing Kong Gateway on Docker and deploying an enterprise
license to a Kong instance, I created includes and inserted the text
in both places from a single doc. I also made the headings more
informative and updated the admonitions to the new style. And I
rebased the branch with all of the updates over the last few weeks.