docs(oss-2.5.x) regenerate autodocs for 2.5
The Admin API now includes both the db and db-less modules in a single
one. These changes include:

* Putting declarative and db-less sections at the top
* Marking the methods that are not compatible with DB-less as such

The "movement" at the top of the file is due to a new version of
openresty and cannot be avoided easily. It is recommended to review
the diff with "ignoring blank space differences" as some end-of-line
spaces were removed.