docs(oss-2.5.x) regenerate autodocs for 2.5 (#2951)
* docs(oss-2.5.x) regenerate autodocs for 2.5

The Admin API now includes both the db and db-less modules in a single
one. These changes include:

* Putting declarative and db-less sections at the top
* Marking the methods that are not compatible with DB-less as such

The "movement" at the top of the file is due to a new version of
openresty and cannot be avoided easily. It is recommended to review
the diff with "ignoring blank space differences" as some end-of-line
spaces were removed.

* Addes db-less compatible badge to admin api

* docs(autodocs) fix navigation

* docs(autodoc) include dbless badges generated from autodocs

* Moves badges to be inline with headings by using markdown formatting

* use new syntax for dbless badges in generator

* Changes badge format to html instead of markdown

* docs(autodoc) generate title badges cleanly

Co-authored-by: Heather Cloward <heathercloward@gmail.com>