doc(kic) use kube-prometheus-stack chart
Rework the KIC Prometheus+Grafana guide to use
https://github.com/prometheus-community/helm-charts/tree/main/charts/kube-prometheus-stack

This is an umbrella chart for Grafana, Prometheus (using the Prometheus
Operator), and other common related tooling. We need to use the Operator
for compatibility with our modern example Prometheus configuration, and
there isn't any obvious reason to continue recommending separate
Prometheus and Grafana charts.