use code 308 instead of 301 for https redirects
Per the Mozilla Developer Web Docs on http 308 "A browser redirects to this page and search engines update their links to the resource ... The request method and the body will not be altered, whereas 301 may incorrectly sometimes be changed to a GET method."

curl and other utilities that rely on it (including Postman) exhibit this behavior when they receive http 301 in response to a POST request. I suggest changing this documentation to use code 308 so that clients are still automatically redirected (does not happen with http 426) and the client knows to preserve the original http verb when making its subsequent request to the URL given by the `Location` response header.