Update docker.md
Added DB less example for EE edition.  Also no mention here about license setup, or the differences between docker on OSX, as this looks like it's written from the linux point of view? Also maybe we need the same for windows? ( I do not have a windows system so I cannot test that ).