Update using-kongclusterplugin-resource.md (#3899)
* Update using-kongclusterplugin-resource.md

Removed the '$' which prevented easy cut and paste, differentiated between command and results (  wonder if there is a better way to do that ).  Also an interesting thing I use a mac which defaults to zsh now, unfortunately I wanted to comment the commands but zsh in native mode does not allow comments in the interactive shell ( that was news to me. ) so wonder if we should be recommending `bash` as the execution environment - then we could use comments to aid documentation. Additionally if people like me save terminal session logs the comments will be there to when we cut and paste the commands into a term.

I could not find the template for commands by numbers that seem to appear in some pages, that may also be a thing to include to show the steps?

* KongClusterPlugin guide results cleanup

Co-authored-by: Michael Heap <m@michaelheap.com>