Removing OS part of image tag names
It seems we are moving away from Alpine image recommendations (for prod systems) in 3.0 to Debian or RHEL based images instead. As a result of that change in 3.0, our docs should reflect that better in the commands we provide for installing an instance of Kong in Docker, since we already called it out as not production-worthy in the main Install Kong page. 

Since this doc is generated based on the selected version of Kong though, I felt it was perhaps best to remove all references to alpine in the tags and stick to just the version itself. Reason behind that was that Alpine is the default image in 2.x when using just the version tag, and Debian is the default image for 3.x when using just the version tag, achieving the appropriate image type already by using just the version tag alone.