Request to modify the OIDC plug in documentation
Okta configuration by default send `openid` scope under the claim label `scp`. By default in the plug the label is set to 'scope' and therefore developers have to spend time trying to figure out why the default scopes are not being passed. This documentation correct aims to reduce the friction for developers.