pr/okta-oidc-doc-changes (#4593)
* Request to modify the OIDC plug in documentation

Okta configuration by default send `openid` scope under the claim label `scp`. By default in the plug the label is set to 'scope' and therefore developers have to spend time trying to figure out why the default scopes are not being passed. This documentation correct aims to reduce the friction for developers.

* Changes to admin API call for OpenID Plugin

I don't believe we have to make changes to the API call under `Consumer Mapping` but there was a change required under `Configure plugin with Admin API`. I have made that change. Now the action done during the UI set up are the same as the admin API.

* Apply suggestions from code review

Co-authored-by: Angel <Guaris@users.noreply.github.com>