Make plugin pages work with --watch and incremental builds
Before this change, Jekyll was re-generating every plugin page on any
file change.
The way Jekyll decides whether to regenerate a page or not is
by checking:
  * whether the source file changed
  * one of its dependencies changed, i.e. layout and includes

First, it checks whether its dependencies changed and if not, it checks
whether the source file changed by checking its mtime,
see
https://github.com/jekyll/jekyll/blob/d78c65f2c7edb6c9b864a0095f81c12d5e075ae6/lib/jekyll/regenerator.rb#L179-L192.
However, first it checks if the source file exists, and the path that it
uses is `document.relative_path`, see
https://github.com/jekyll/jekyll/blob/d78c65f2c7edb6c9b864a0095f81c12d5e075ae6/lib/jekyll/regenerator.rb#L168.

If the `relative_path` is not set, it uses `@dir` to build it see
https://github.com/jekyll/jekyll/blob/d78c65f2c7edb6c9b864a0095f81c12d5e075ae6/lib/jekyll/page.rb#L144-L146
Given that we set `@dir` and `@name`, the `relative_path` ended up
pointing to a File that didn't exist. Hence, the check failed causing
the page to be regenerated.