Makes the PDK file structure slightly more consistent
The table at the bottom of the page lists filenames the plugin developer should use
And the descriptions below go into more detail on each file
The table and the descriptions are inconsistent regarding the API interface filename

I looked at existing plugins, and there seems to be a standard to use api.lua and not admin-api.lua
despite that being the name of the referenced autodoc file.

I suggest we make the filename in the description match the filename in the table so the
developer knows what to name their file.