Add note about seeding super-admin password
A customer highlighted the concern that if following the documentation in-order as currently defined then it's too late to set the Super-Admin password by the time that section is read. Added a quick note to encourage Enterprise users to skip to the Seeding Super-Admin password section instead of running the initial ```kong migrations``` command without it.