Update description for rediscovery_lifetime field
Updated the description of config.rediscovery_lifetime based on feedback from Kong Support in current case.
For visibility, feedback was as follows:

Kong only tries to rediscover once a token comes in for which it doesn't have a key in the cache. If you have set config.rediscovery_lifetime to 300 (5 mins), which will just determine how long Kong waits to discover the keys again. So if a token comes in that has been signed with a key that a KID kong doesn't have in its cache, it will rediscover the keys. If another one comes in after that, which again has a new KID in less than 300s, Kong will not rediscover.