Update index.md to include details about JWK cache duration (#3719)
* Update index.md

Documentation should indicate how long discovery information remains cached from the created_at time.
Note to Kong engineer, please update PR with required time period value or provide relevant config value name for cache duration.

* Update based on details provided by Kong Support

Was tagged in a case where the Kong Support member advised on the cache refresh behaviour. Updated document changes accordingly

* Update app/_hub/kong-inc/openid-connect/index.md

* Update description for rediscovery_lifetime field

Updated the description of config.rediscovery_lifetime based on feedback from Kong Support in current case.
For visibility, feedback was as follows:

Kong only tries to rediscover once a token comes in for which it doesn't have a key in the cache. If you have set config.rediscovery_lifetime to 300 (5 mins), which will just determine how long Kong waits to discover the keys again. So if a token comes in that has been signed with a key that a KID kong doesn't have in its cache, it will rediscover the keys. If another one comes in after that, which again has a new KID in less than 300s, Kong will not rediscover.

* Update app/_hub/kong-inc/openid-connect/index.md

Co-authored-by: Aapo Talvensaari <aapo.talvensaari@gmail.com>

* Add discovery cache details

Updated details around the behavior of the discovery cache.

* Fix plugin config

* Appease Vale, fix broken links

Signed-off-by: Diana <75819066+cloudjumpercat@users.noreply.github.com>

---------

Signed-off-by: Diana <75819066+cloudjumpercat@users.noreply.github.com>
Co-authored-by: lena-larionova <54370747+lena-larionova@users.noreply.github.com>
Co-authored-by: Aapo Talvensaari <aapo.talvensaari@gmail.com>
Co-authored-by: Fabian Rodriguez <fabian.rodriguez@konghq.com>
Co-authored-by: Diana <75819066+cloudjumpercat@users.noreply.github.com>