Support policy tab (#5374)
* remove mesh 1.0.x and 1.1.x

These features have been out of support for a long time and
don't include the automation we leverage for newer releases.

Therefore remove these as it's simpler

Signed-off-by: Charly Molter <charly.molter@konghq.com>

* feat(jekyll): add tag to generate the support policy table

The Mesh team is generating their versions.yml from metadata and syncing it in this repo.
This adds the following features to leverage this:

- The policy table is dynamically generated from the versions.yml
- Add the possibility to have a recent version that isn't released yet by having the latest pointer be explicit
- We can explicitly ignore some release from rendering (useful for dev for example)

Signed-off-by: Charly Molter <charly.molter@konghq.com>

---------

Signed-off-by: Charly Molter <charly.molter@konghq.com>