Update systemd.md (#5669)
* Update systemd.md

I got a common error with customers who are attempting to run with the default systemd file. The issue is that we need to add on the user and group as root (2 additional lines) to run the system file. I have added an example here. 

The next question of how to run as non-root will be raised. Understand that we do have a doc here: https://docs.konghq.com/gateway/latest/production/running-kong/kong-user/, but the important note is sending mixed signal that Kong can be only run as root. Requesting for engineering team to work on a tested solution to run Kong without root - that will be more palatable to customers who are  concerned with security

* Change header to explanation

---------

Co-authored-by: lena-larionova <54370747+lena-larionova@users.noreply.github.com>