Clarify group claims / team mapping in Konnect Okta setup
Clarified that the claim set up is optional as group claims are only needed if the admin intends to use Konnect team mapping functionality. It was marked optional in one sentence, but when looking at the later part it implied it's a requirement, so attempted to clarify this in the docs.