Update enable-app-reg.md
Updating w/ feedback from Ross: 

https://deploy-preview-5760--kongdocs.netlify.app/konnect/dev-portal/applications/enable-app-reg/
This phrasing seems awkward: "We are rolling out full support in any non-default runtime group, using the konnect-application-auth plugin that was created for Kong Gateway 3.0." Specifically "rollout out" -- we already support non-default runtime groups.  We shoudl jsut be describing how the behaivor is different
See default runtime group for additional context. --> Default runtime group link is dead
Note: It is recommended but not required for an API product version is linked to a Gateway service. --> It is required for app reg to work, however.