Fix performance issue caused by plugin pages without `source_file`
We use the `source_file` as the page's `relative_path` which is used
internally by Jekyll to konw which pages should be re-generated whenver
a file changes.

Configuration and plugin hub example pages were setting it to `nil`
causing all those pages to be re-generated whenever any file changed.
This fixes the issue by setting the right `source_file` to those pages.