Remove contradictory line on upgrade latest page
Remove contradictory statement about lowest version for direct upgrade to 3.4 being 3.0, when the table later on the page suggests otherwise. Since the table is the "source of truth" for the details, no need to keep a line earlier that would always need to be manually updated to match the table.