Updated RLA index page to share some details from RL plugin
There were some missing details from the RL plugin which I pasted into the RLA plugin index page. Also added a section for how the IP address is determined as this was missing from the details and was not obvious to customers how this worked when limiting by IP address.