Clarify and improve DP Resilience feature notes
In troubleshooting a support case, we discovered that there was no mention of the `AWS_SESSION_TOKEN` variable which is needed in certain situations. Added a line in the Configuration section to clarify this for AWS S3.

Some customers were also confused thinking they needed to manually create the file structure including the config.json file in the bucket, so I added a line that confirmed this is created automatically and should not be created manually.

Also aligned some wording between AWS S3 and GCP bucket instructions, just minor tweaks to verbiage.