Sectioned cluster_fallback configs and added WebAssembly text to Wasm section
Moved the DP Resilience feature (Cluster Fallback) properties to their own section for better organization and linking. 

Also noticed our Wasm section had no reference to the text 'WebAssembly' which resulted in it not being found easily. Added that to the title of that section to allow for better searchability.