fix(docs): remove the influxdb settings when starting Kong gateway as it's not available yet
Remove the influxdb related environment variables when starting the kong gateway as at the point of the doc the influxdb instance hasn't been started yet (It's started at a later step in the doc).

Another option to correct this is to move the steps of starting an influxdb instance before the gateway start steps. This requires some rewording of the steps so I'd leave the change to the docs team.