Sectioned cluster_fallback configs + added WebAssembly text to Wasm section (#6065)
* Sectioned cluster_fallback configs and added WebAssembly text to Wasm section

Moved the DP Resilience feature (Cluster Fallback) properties to their own section for better organization and linking. 

Also noticed our Wasm section had no reference to the text 'WebAssembly' which resulted in it not being found easily. Added that to the title of that section to allow for better searchability.

* Re-worked the ordering for cluster_fallback_* properties

Resolving the suggestion from @lena-larionova that the ordering should match the kong.conf file.

* add changes from source file

---------

Co-authored-by: lena.larionova <yelena.larionova@gmail.com>