Update _index.md
Based on questions in slack, the example in docs does not mention how the plugin works with the access token specified as query string. The plugin uses the default of `access_token` as. the query string parameter name. This is not specified anywhere. Hence this example shows how to use the plugin with the query string param as option instead of header