DOCU-3435: Find in page counts hidden elements in the result
This was caused due to the way the sidebar items (accordion items) expanded and collapsed.
When an accordion item was "collapsed", its:

max-height was set to 0
opacity was set to 1 so technically even though they were "hidden" to the naked eye, they were still "visible".
This caused Chrome's and Firefox's find in page to count the supposedly "hidden" occurrences of the searched text as part of the results.

Setting the visibility to hidden when the item is collapsed, and to visible when it is expanded fixed the issue.