fix(opentelemetry): address invalid request to set up plugin
The request documented to configure the Opentelemetry plugin is invalid
because it uses a dot `.` in the name of the attribute, which is
interpreted as a field separator in the curl form data.

This updates the documentation without changing the example, using JSON
instead of form data to allow passing a name that includes a dot.