Add optional filters to search page so that gateway comes first
This shouldn't be here. I'm waiting a response from support so that we
can use a Rule that adds optional filters based on the query, e.g. if a
query contains `gateway` we would add an optional filter that sets a
higher score to `Kong Gateway` records. Same for Mesh, etc.