Update: OIDC plugin example for using an access token (#6231)
* Update _index.md

Based on questions in slack, the example in docs does not mention how the plugin works with the access token specified as query string. The plugin uses the default of `access_token` as. the query string parameter name. This is not specified anywhere. Hence this example shows how to use the plugin with the query string param as option instead of header

* set the list of IdPs to be ignored by vale, they're all company names

* Convert to cURL and minor formatting

* add NATs to dictionary

* fix curl command based on feedback

---------

Co-authored-by: lena-larionova <54370747+lena-larionova@users.noreply.github.com>