Self-host Google Fonts Stylesheet
To enhance the load performance in certain scenarios, we bundle
the Roboto fonts stylesheet in app assets instead of fetching it
from `fonts.googleapis.com` directly.

This patch is part of the DOCU-3037

The font file hosted on `fonts.gstatic.com` is guaranteed to be
existed across the fonts version update so it's safe to self-host
a snapshot of the stylesheet.