[DOCU-3662]: Improve plugin versioning by adding minimum_version and maximum_version support to frontmatters
Before this change, whenever we added a new page for a specific version
we had to copy the entire's plugin directory and move it into a new
folder.

This change adds support for `minimum_version` and `maximum_version` in
the frontmatter of plugin pages.

* If a page has one of these set in frontmatter, it gets generated
for that version range, similar to versions.yml
* If a page doesn’t have any minimum or maximum version,
it defaults to the setting for the whole plugin, which is set in versions.yml

The only caveat is that we had to replace `_metadata.yml` with a specific
`_metadata` folder. If a file for a specific version exists, (e.g.
`_metadata/_2.6.x.yml`) it will be used for that specific release, if
not `_metadata/_index.yml` will be used.