[Plugin Hub] Translate sidebar using string translations
We'll show the sidebar in japanese if the docs_nav for the corresponding
gateway release is available in japanese. Otherwise, we'll show the
english version.

This approach is slightly different to the one we take with the docs,
where we display the sidebar in japanese if both the docs_nav file and
the source file are available in japanese.
I'll try to take the same approach for consistency, but sidebars are
built differently, so it's tricky.