[Plugin Hub] Display deprecated fields at the bottom of the schema
As of version 3.7.x, fields in the plugin schema have been flagged with
a `deprecation` key, containing information about the deprecation.

This commit adds the fields marked as deprecated to the list of
`Deprecated Parameters` on the Configuration Reference page.
It also renders a deprecation notice, with the information provided in
the field.

Note: Fields marked as deprecated are not rendered in the main body of
the schema, just in the `Deprecated Parameters` section.