Fixes check-links.yml The broken link checker didn't catch a broken link when it was run in "PR" mode, but the full scan did.
Related PR: https://github.com/Kong/docs.konghq.com/pull/7381

In "PR" mode, the check pull the modified files, gets the corresponding
urls, and checks those urls for broken links.

The problem was with the filenames-to-URLs transformation.
The way it works is by checking the `docs_nav` files for an entry that has a file as
`src` or `url`. If an entry in a nav file didn't have a `src` we used
the `url` as `src`. See
https://github.com/Kong/docs.konghq.com/blob/main/tools/_utilities/lib/nav-to-urls.js#L34-L38.

However, this didn't consider the case where the `url` corresponds to a
file that ends with `index.md`

Example:
In a PR `app/_src/gateway/ai-gateway/index.md` was modified and a broken
link is introduced.

The corresponding entry in gateway's `doc_nav` file is:
```
- text: Overview
  url: /ai-gateway/
```

It has no `src`, so the `url` is used to find the corresponding source
file during single source generation. Single Source Generation checks
for `app/_src/gateway/ai-gateway.md` and `app/_src/gateway/ai-gateway/index.md`.

However, the broken link checker in "PR" mode only checked for
`app/_src/gateway/ai-gateway.md`. Hence, it couldn't find the file and
didn't generate a url to check.

This commit fixes the issue by checking for the existence of both files.