Use jekyll_plugins to group gems in the Gemfile This way they get required by Jekyll on build
Fixes the build in Ubuntu by loading the plugins correctly before
running the code. The problem was with the order the plugins were being loaded.
Which used to work because jekyll-generator-single source was being
manually required, but that was removed as part of
https://github.com/Kong/docs.konghq.com/pull/7395.

Even without that required, it used to work because of the`priorities`.
The jekyll-generator-single-source had the highets and that's why it was
being loaded first. but now priorities changed so that no longer
applies.

Using the jekyll_plugins group in the Gemfile ensures that they are
loaded.