[DOCU-3873] Handle canonical urls.
According to the following rules:
* All /latest/ or evergreen URLs (i.e plugin hub) should have self referencing canonicals 
* All /{version #}/ URLs should have a canonical that points to the /latest/
* Canonical urls in the JP site should point to  docs.jp.konghq.com
* Canonical urls in  docs.konghq.com, point to  http://docs.konghq.com

Context: Why not set the canonical to the default docs LP if the page is not translated?
The "JP" in the subdomain is a country code + the hreflang shows that everything under this subdomain structure is meant for Japan search results.
If we pointed the canonical to the default docs LP then we are essentially telling search engines that we do not want this content to show up in Japan.
Submitting a bulk number of URLs to search engines and then telling it not to index a chunk of those is confusing and can hurt us from a holistic level.
So even though it is in english, it is best to still leverage the jp subdomain structure to avoid indexing tension.