chore(rate-limiting): clarify rate-limiting accuracy
Both rate-limiting and rate-limiting-advanced plugins documentation specify
that various policies are possible and states that cluster and redis are
"Accurate". However it's possible to configure a `sync_rate` option
for these plugins and set it to a value that does not offer synchronous
behaviour allowing for some slippage in favor of performance.

This commit clarifies that `sync_rate` affects mentioned accuracy.

KAG-2896