Run smoke tests with JEKYLL_ENV=production so that the right config gets picked up and the sitemap gets generated
The previous step builds the docs using JEKYLL_ENV=production, but
running netlify-dev, re-run the site using JEKYLL_ENV=development which
deletes the sitemap pages, because we use a generator to build those.

This didn't happen before because the sitemap was generated differently.
There was a file app/sitemap.xml, so running netlify dev would serve
this page from the cache.