Clarify workspace instead of environment
According to tests, this seems to be unique to each workspace. In other words, one workspace can still use dynamic ordering if consumer-scoped plugins are in use as long as the consumer-scoped plugins are in a different workspace than the one using dynamic ordering for example. This created confusion with a customer and required manual testing.