Update _index.md
Updated the FAQ to include a note about how the certificate should  be passed. The docs were not entirely clear that base64 and url encoding require different data to be passed (one including the being/end certificate delimiters). 

If this is not the proper place for the details I can update as needed.